requirements.txt
torch
opencv-python
numpy
Pillow
tensorflow
matplotlib
tflite-runtime